Support 64-bit lengths in combineCRC32
This is basically a revert of 3c0cd2e03ec9c8b0223ed21f71321af75bc1b938.

Saving 2 KiB from a global variable that's only initialised once isn't
really worth it. This way we can support any length for precompressed
data.

name                old time/op    new time/op    delta
PrecomputeCRC32-12     104µs ± 1%     147µs ± 2%  +41.10%  (p=0.000 n=10+10)
CombineCRC32-12        763ns ± 2%     779ns ± 2%   +2.02%  (p=0.003 n=10+10)

name                old alloc/op   new alloc/op   delta
PrecomputeCRC32-12    6.14kB ± 0%    8.19kB ± 0%  +33.33%  (p=0.000 n=10+10)
CombineCRC32-12        0.00B          0.00B          ~     (all equal)

name                old allocs/op  new allocs/op  delta
PrecomputeCRC32-12      1.00 ± 0%      1.00 ± 0%     ~     (all equal)
CombineCRC32-12         0.00           0.00          ~     (all equal)

+++ builder.go
@@ -208,7 +208,7 @@ func (b *builder) AddPrecompressedData(data *PrecompressedData) {
 
 	if !b.rawDeflate {
 		b.size += uint32(data.size)
-		b.crc = combineCRC32(crc32Mat, b.crc, data.crc, uint64(data.size))
+		b.crc = combineCRC32(crc32Mat, b.crc, data.crc, data.size)
 	}
 
 	_, b.err = b.w.Write(data.bytes)
@@ -457,7 +457,7 @@ type PrecompressedData struct {
 	level int
 
 	bytes []byte
-	size  int64
+	size  uint64
 	crc   uint32
 }
 
@@ -479,7 +479,7 @@ type PrecompressedWriter struct {
 	buf *bytes.Buffer
 	fw  *flate.Writer
 
-	size int64
+	size uint64
 	crc  uint32
 
 	lastFlush bool
@@ -527,7 +527,7 @@ func (w *PrecompressedWriter) Write(p []byte) (int, error) {
 
 	w.lastFlush = false
 
-	w.size += int64(len(p))
+	w.size += uint64(len(p))
 	w.crc = crc32.Update(w.crc, crc32.IEEETable, p)
 
 	n, err := w.fw.Write(p)
